@@ -5,92 +5,136 @@
   "dependencies": {
     "@albertli/openfin-browser-adapter": "^0.10.5",
     "@albertli/redux-openfin": "^0.10.5",
-    "@material-ui/core": "^1.5.0",
-    "@material-ui/icons": "^2.0.2",
+    "@babel/core": "7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/plugin-proposal-decorators": "^7.0.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/plugin-transform-classes": "^7.0.0",
+    "@babel/plugin-transform-destructuring": "^7.0.0",
+    "@babel/plugin-transform-modules-commonjs": "^7.0.0",
+    "@babel/plugin-transform-react-constant-elements": "^7.0.0",
+    "@babel/plugin-transform-react-display-name": "^7.0.0",
+    "@babel/plugin-transform-regenerator": "^7.0.0",
+    "@babel/plugin-transform-runtime": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-flow": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/preset-typescript": "^7.0.0",
+    "@babel/runtime": "7.0.0",
+    "@material-ui/core": "^3.1.1",
+    "@material-ui/icons": "^3.0.1",
     "@types/classnames": "^2.2.6",
     "@types/enzyme": "^3.1.13",
+    "@types/enzyme-adapter-react-16": "^1.0.3",
     "@types/history": "^4.7.0",
     "@types/jest": "^23.3.1",
     "@types/node": "^10.7.0",
-    "@types/react": "^16.4.10",
+    "@types/react": "^16.4.13",
     "@types/react-dom": "^16.0.7",
-    "@types/react-redux": "^6.0.6",
     "@types/react-router-dom": "^4.3.0",
     "@types/redux-actions": "^2.3.0",
-    "autoprefixer": "7.1.6",
-    "babel-core": "6.26.0",
-    "babel-eslint": "7.2.3",
-    "babel-jest": "20.0.3",
-    "babel-loader": "7.1.2",
-    "babel-preset-react-app": "^3.1.1",
-    "babel-runtime": "6.26.0",
-    "case-sensitive-paths-webpack-plugin": "2.1.1",
-    "chalk": "1.1.3",
+    "autoprefixer": "8.5.0",
+    "babel-core": "7.0.0-bridge.0",
+    "babel-eslint": "8.2.3",
+    "babel-jest": "22.4.3",
+    "babel-loader": "8.0.0-beta.4",
+    "babel-plugin-named-asset-import": "1.0.0-next.a671462c",
+    "babel-preset-react-app": "4.0.0-next.a671462c",
+    "bfj": "5.2.0",
+    "case-sensitive-paths-webpack-plugin": "2.1.2",
+    "chalk": "2.4.1",
     "classnames": "^2.2.6",
     "color": "^3.0.0",
-    "css-loader": "0.28.7",
-    "dotenv": "4.0.0",
+    "css-loader": "0.28.11",
+    "dotenv": "5.0.1",
     "dotenv-expand": "4.2.0",
-    "enzyme": "^3.4.1",
-    "enzyme-adapter-react-16": "^1.2.0",
-    "eslint": "4.10.0",
-    "eslint-config-react-app": "^2.1.0",
-    "eslint-loader": "1.9.0",
-    "eslint-plugin-flowtype": "2.39.1",
-    "eslint-plugin-import": "2.8.0",
-    "eslint-plugin-jsx-a11y": "5.1.1",
-    "eslint-plugin-react": "7.4.0",
+    "enzyme": "^3.5.0",
+    "enzyme-adapter-react-16": "^1.3.1",
+    "enzyme-to-json": "^3.3.4",
+    "eslint": "4.19.1",
+    "eslint-config-react-app": "3.0.0-next.a671462c",
+    "eslint-loader": "2.0.0",
+    "eslint-plugin-flowtype": "2.46.3",
+    "eslint-plugin-import": "2.11.0",
+    "eslint-plugin-jsx-a11y": "6.0.3",
+    "eslint-plugin-react": "7.8.2",
     "express": "^4.16.3",
-    "extract-text-webpack-plugin": "3.0.2",
-    "file-loader": "1.1.5",
-    "fs-extra": "3.0.1",
+    "file-loader": "1.1.11",
+    "fork-ts-checker-webpack-plugin": "^0.4.9",
+    "fs-extra": "5.0.0",
+    "graphql": "0.13.2",
+    "graphql-tag": "2.9.2",
     "history": "^4.7.2",
-    "html-webpack-plugin": "2.29.0",
-    "jest": "21.2.1",
+    "html-webpack-plugin": "3.2.0",
+    "identity-obj-proxy": "3.0.0",
+    "jest": "^23.6.0",
+    "loader-utils": "^1.1.0",
     "material-ui-pickers": "^1.0.0-rc.13",
+    "mini-css-extract-plugin": "^0.4.0",
     "moment": "^2.22.2",
     "morgan": "^1.9.0",
     "object-assign": "4.1.1",
     "openfin-launcher": "^1.4.4",
-    "postcss-flexbugs-fixes": "3.2.0",
-    "postcss-loader": "2.0.8",
+    "optimize-css-assets-webpack-plugin": "^4.0.1",
+    "postcss-flexbugs-fixes": "3.3.1",
+    "postcss-loader": "2.1.5",
     "promise": "8.0.1",
     "raf": "3.4.0",
-    "react": "^16.4.2",
+    "react": "^16.5.0",
     "react-custom-scrollbars": "^4.2.1",
-    "react-dev-utils": "^5.0.1",
-    "react-dom": "^16.4.2",
+    "react-dev-utils": "6.0.0-next.a671462c",
+    "react-dom": "^16.5.0",
     "react-json-view": "^1.19.1",
-    "react-number-format": "^3.5.1",
+    "react-number-format": "^3.6.0",
     "react-particles-js": "^2.3.0",
     "react-redux": "^5.0.7",
     "react-router-dom": "^4.3.1",
     "react-splitter-layout": "^3.0.1",
     "redux": "^4.0.0",
     "redux-actions": "^2.6.1",
+    "redux-mock-store": "^1.5.3",
     "redux-saga": "^0.16.0",
     "resolve": "1.6.0",
     "rimraf": "^2.6.2",
+    "sass-loader": "7.0.1",
     "shortid": "^2.2.13",
-    "source-map-loader": "^0.2.3",
-    "style-loader": "0.19.0",
-    "sw-precache-webpack-plugin": "0.11.4",
-    "ts-jest": "^23.1.3",
-    "ts-loader": "^3.5.0",
-    "typescript": "^3.0.1",
-    "url-loader": "0.6.2",
-    "webpack": "3.8.1",
-    "webpack-dev-server": "2.9.4",
-    "webpack-manifest-plugin": "1.3.2",
-    "whatwg-fetch": "2.0.3"
+    "style-loader": "0.21.0",
+    "svgr": "1.9.2",
+    "sw-precache-webpack-plugin": "0.11.5",
+    "tcp-port-used": "^1.0.0",
+    "thread-loader": "1.1.5",
+    "ts-jest": "^23.1.4",
+    "tslint": "^5.11.0",
+    "tslint-config-prettier": "^1.15.0",
+    "tslint-react": "^3.6.0",
+    "typescript": "^3.0.3",
+    "uglifyjs-webpack-plugin": "1.2.5",
+    "url-loader": "1.0.1",
+    "webpack": "4.8.3",
+    "webpack-dev-server": "3.1.5",
+    "webpack-manifest-plugin": "2.0.3",
+    "whatwg-fetch": "2.0.4"
   },
   "scripts": {
     "start": "node scripts/start.js",
     "start-openfin": "node scripts/start.openfin.js",
     "build": "rimraf build && node scripts/build.js",
-    "server": "node scripts/server.js",
-    "server-openfin": "node scripts/server.openfin.js",
+    "serve": "node scripts/server.js",
+    "serve-openfin": "node scripts/server.openfin.js",
     "test": "node scripts/test.js --env=jsdom"
+  },
+  "browserslist": {
+    "development": [
+      "last 2 chrome versions",
+      "last 2 firefox versions",
+      "last 2 edge versions"
+    ],
+    "production": [
+      ">0.25%",
+      "not op_mini all",
+      "ie 11"
+    ]
   },
   "jest": {
     "collectCoverageFrom": [
