@@ -1,4 +1,4 @@
-import Enzyme from 'enzyme';
+import Enzyme from'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 
 Enzyme.configure({adapter:new Adapter()});
