@@ -16,6 +16,19 @@ const app = express();
 app.use(logger('dev'));
 app.use(express.static(paths.appBuild));
 
+app.use((req,res,next)=>{
+    if (req.accepts('html')){
+        res.status(200);
+        res.sendFile(paths.appBuild+'/index.html');
+    }else{
+        res.status(404);
+        if(req.accepts('json')){
+            res.send({error:'Not found'});
+        }
+        res.type('txt'),send('Not found');
+    }
+});
+
 var port = normalizePort(process.env.PORT || '3000');
 app.set('port',port);
 
@@ -67,8 +80,8 @@ function onError(error){
 function onListening() {
     var addr = server.address();
     var bind = typeof addr === 'string'
-        ? 'pipe' + addr
-        : 'port' + addr.port;
+        ? 'pipe ' + addr
+        : 'port ' + addr.port;
     debug('Listening on ' + bind);
     console.log('Listening on ' + bind);
 }
