@@ -1,5 +1,5 @@
 import {
-    grayColor,
+    greyColor,
     primaryColor,
     infoColor,
     successColor,
@@ -12,7 +12,7 @@ const buttonStyle = {
     button: {
         minHeight: "auto",
         minWidth: "auto",
-        backgroundColor: grayColor,
+        backgroundColor: greyColor,
         color: "#FFFFFF",
         boxShadow:
             "0 2px 2px 0 rgba(153, 153, 153, 0.14), 0 3px 1px -2px rgba(153, 153, 153, 0.2), 0 1px 5px 0 rgba(153, 153, 153, 0.12)",
@@ -36,7 +36,7 @@ const buttonStyle = {
         cursor: "pointer",
         "&:hover,&:focus": {
             color: "#FFFFFF",
-            backgroundColor: grayColor,
+            backgroundColor: greyColor,
             boxShadow:
                 "0 14px 26px -12px rgba(153, 153, 153, 0.42), 0 4px 23px 0px rgba(0, 0, 0, 0.12), 0 8px 10px -5px rgba(153, 153, 153, 0.2)"
         },
@@ -76,7 +76,7 @@ const buttonStyle = {
     white: {
         "&,&:focus,&:hover": {
             backgroundColor: "#FFFFFF",
-            color: grayColor
+            color: greyColor
         }
     },
     rose: {
