@@ -1,5 +1,5 @@
 import * as React from 'react';
-import * as cx from 'classnames';
+import cx from 'classnames';
 
 import { withStyles } from '@material-ui/core';
 import Button from '@material-ui/core/Button';
